requirements.txt
absl-py==0.7.1
argon2-cffi==20.1.0
astor==0.8.0
attrs==20.2.0
backcall==0.2.0
bleach==3.1.5
certifi==2016.2.28
cffi==1.14.2
cloudpickle==1.2.1
contextlib2==0.5.5
Cython==0.29.21
decorator==4.4.0
defusedxml==0.6.0
dm-sonnet==1.34
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
Keras==2.1.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.16.4
olefile==0.46
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==4.2.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.9.1
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.1.2
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
scikit-learn==0.23.1
scipy==1.3.0
semantic-version==2.6.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
tensorboard==1.14.0
tensorboardX==1.8
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.4.0
torchvision==0.6.0a0+82fd1c8
tornado==6.0.4
tqdm==4.32.2
traitlets==4.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
